+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 -   Fix angular z-index system (revert to previous system)
+-   Fix `Lightbox` close on escape.
 
 ## [1.0.1][] - 2021-01-13